test(context-optimizer): split at the measurement/classification seam
The file reached 507 lines and tripped the taste file-size rule, taking the tree
to 603 against a baseline of 602. main carries it at 490, so this branch pushed
it over; trimming the two comments I had just added got it to 505, which is still
over and would re-cross on the next edit.

Split on the seam the module already has: measurement primitives (content,
tool calls, action verbs, reference ratio, trigger patterns) stay, and everything
built on them (classification, hybrid recommendations, analyze_content, CLI)
moves out. 349 and 240 lines.

111 tests pass across the directory, mypy reports Success on both files,
taste-lints reports 0 errors on each, and the count ratchet passes.
